feat: Move create follow up form to a separeted modal
